refactor: the extra folder app was removed
