+++ src/lib/constants/routes.const.ts
@@ -1,11 +1,10 @@
 export const ROUTES = {
-  HOME: '/',
-  APP: '/app',
-  APP_HELP: '/app/help',
-  APP_PROFILE: '/app/profile',
-  APP_CREATE_PROFILE: '/app/create-profile',
-  APP_CREATE_ADDRESS: '/app/create-address',
-  APP_SELECT_ENTITY: '/app/select-entity',
+  APP: '/',
+  APP_HELP: '/help',
+  APP_PROFILE: '/profile',
+  APP_CREATE_PROFILE: '/create-profile',
+  APP_CREATE_ADDRESS: '/create-address',
+  APP_SELECT_ENTITY: '/select-entity',
 } as const;
 
 export type ObjectValues<T> = T[keyof T];
